Roll Skia DEPS to r7869
control: https://codereview.chromium.org/12335104/
Review URL: https://codereview.chromium.org/12319120

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@184947 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 5264cde7cec1460e0c0e540829d1a6456639eed1

+++ skia/ext/platform_canvas_unittest.cc
@@ -413,7 +413,7 @@ TEST(PlatformBitmapTest, PlatformBitmap) {
   EXPECT_FALSE(platform_bitmap->GetBitmap().isNull());
   EXPECT_EQ(kWidth, platform_bitmap->GetBitmap().width());
   EXPECT_EQ(kHeight, platform_bitmap->GetBitmap().height());
-  EXPECT_LE(platform_bitmap->GetBitmap().width()*4,
+  EXPECT_LE(static_cast<size_t>(platform_bitmap->GetBitmap().width()*4),
             platform_bitmap->GetBitmap().rowBytes());
   EXPECT_EQ(SkBitmap::kARGB_8888_Config,  // Same for all platforms.
             platform_bitmap->GetBitmap().config());
